requirements.txt
fabric
termcolor
arrow
pyyaml
prettytable